git
Lane Line detection is a critical component for self driving cars and also for computer vision in general. This concept is used to describe the path for self-driving cars and to avoid the risk of getting in another lane. In this article, we will build a machine learning project to detect lane lines in real-time.